requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
PyJWT==2.8.0
sqlalchemy==2.0.23

# Advanced Quality Assurance Dependencies
psutil==5.9.6
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7
sentry-sdk[fastapi]==1.38.0

# Security & Validation
python-multipart==0.0.6
email-validator==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Performance & Caching
redis==5.0.1
cachetools==5.3.2

# Testing & Quality
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring & Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
